+++ utils/wl_generate_acks.py
@@ -146,7 +146,7 @@ ACKS = [
         'MIT', 'https://github.com/explosion/spacy-pkuseg/blob/master/LICENSE'
     ], [
         'Stanza', 'https://github.com/stanfordnlp/stanza',
-        '1.7.0', 'Peng Qi (齐鹏), Yuhao Zhang (张宇浩),<br>Yuhui Zhang (张钰晖), Jason Bolton,<br>Tim Dozat, John Bauer',
+        '1.8.2', 'Peng Qi (齐鹏), Yuhao Zhang (张宇浩),<br>Yuhui Zhang (张钰晖), Jason Bolton,<br>Tim Dozat, John Bauer',
         'Apache-2.0', 'https://github.com/stanfordnlp/stanza/blob/main/LICENSE'
     ], [
         'SudachiPy', 'https://github.com/WorksApplications/sudachi.rs',