+++ CHANGELOG.md
@@ -29,6 +29,8 @@
 - Utils: Add Pyphen's Basque syllable tokenizer
 - Utils: Add PyThaiNLP's Han-solo
 - Utils: Add spaCy's Faroese and Norwegian (Nynorsk) word tokenizers
+- Utils: Add Stanza's Armenian (Classical) sentence tokenizer, word tokenizer, part-of-speech tagger, lemmatizer, and dependency parser
+- Utils: Add Stanza's English (Old) sentence tokenizer, word tokenizer, part-of-speech tagger, lemmatizer, and dependency parser
 - Utils: Add Stanza's Sindhi part-of-speech tagger
 - Utils: Add VADER's sentiment analyzers
 - Work Area: Add Colligation Extractor - Filter results - Node/Colligation length
@@ -77,7 +79,7 @@
 - Dependencies: Upgrade simplemma to 1.0.0
 - Dependencies: Upgrade spaCy to 3.7.5
 - Dependencies: Upgrade spacy-pkuseg to 0.0.33
-- Dependencies: Upgrade Stanza to 1.7.0
+- Dependencies: Upgrade Stanza to 1.8.2
 - Dependencies: Upgrade SudachiPy to 0.6.8
 - Dependencies: Upgrade Underthesea to 6.8.4
 - Dependencies: Upgrade wordcloud to 1.9.3